refactor: update Results
